+++ R/bit64-package.R
@@ -695,9 +695,9 @@
 #'   ramsort ramsortorder repeat.time setattr shellorder shellsort
 #'   shellsortorder still.identical xor
 #' @importFrom graphics barplot par title
-#' @importFrom methods as is
+#' @importFrom methods as is callGeneric
 #' @importFrom stats cor median quantile
-#' @importFrom utils head packageDescription strOptions tail
+#' @importFrom utils head packageDescription strOptions tail getS3method
 #' @export : :.default :.integer64
 #' @export [.integer64 [[.integer64 [[<-.integer64 [<-.integer64
 #' @export %in% %in%.default