+++ tests/testthat/helper.R
@@ -1,5 +1,6 @@
-# TODO(#45): use matrix() directly
 matrix64 = function(x, nrow=1L, ncol=1L, byrow=FALSE) {
+  if (getRversion() >= "4.0.0") return(matrix(as.integer64(x), nrow, ncol, byrow))
+  warning("using matrix64() with R version < 4.0.0")
   x = as.integer64(x)
   if (byrow) {
     dim(x) = c(ncol, nrow)
@@ -11,6 +12,8 @@ matrix64 = function(x, nrow=1L, ncol=1L, byrow=FALSE) {
 }
 
 array64 = function(x, dim) {
+  if (getRversion() >= "4.0.0") return(array(as.integer64(x), dim))
+  warning("using array64() with R version < 4.0.0")
   x = as.integer64(x)
   dim(x) = dim
   x